Separate command scheduling tests. Introduce abstract base for test suites
