+++ server/src/test/java/org/spine3/server/command/AbstractCommandBusTestSuite.java
@@ -1,0 +1,95 @@
+/*
+ * Copyright 2017, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package org.spine3.server.command;
+
+import org.junit.After;
+import org.junit.Before;
+import org.spine3.base.CommandContext;
+import org.spine3.server.event.EventBus;
+import org.spine3.server.storage.memory.InMemoryStorageFactory;
+import org.spine3.test.command.CreateProject;
+import org.spine3.test.command.event.ProjectCreated;
+import org.spine3.testdata.TestEventBusFactory;
+
+import static org.mockito.Mockito.spy;
+import static org.spine3.base.Identifiers.newUuid;
+
+/**
+ * Abstract base for test suites of {@code CommandBus}.
+ *
+ * @author Alexander Yevsyukov
+ */
+@SuppressWarnings("ProtectedField") // OK for brevity of derived tests.
+public abstract class AbstractCommandBusTestSuite {
+
+    protected CommandBus commandBus;
+    protected CommandStore commandStore;
+    protected EventBus eventBus;
+    protected ExecutorCommandScheduler scheduler;
+    protected CreateProjectHandler createProjectHandler;
+    protected TestResponseObserver responseObserver;
+
+    @Before
+    public void setUp() {
+        final InMemoryStorageFactory storageFactory = InMemoryStorageFactory.getInstance();
+        commandStore = spy(new CommandStore(storageFactory.createCommandStorage()));
+        scheduler = spy(new ExecutorCommandScheduler());
+        commandBus = CommandBus.newBuilder()
+                               .setCommandStore(commandStore)
+                               .setCommandScheduler(scheduler)
+                               .setThreadSpawnAllowed(true)
+                               .setAutoReschedule(false)
+                               .build();
+        eventBus = TestEventBusFactory.create(storageFactory);
+        createProjectHandler = new CreateProjectHandler(newUuid());
+        responseObserver = new TestResponseObserver();
+    }
+
+    @After
+    public void tearDown() throws Exception {
+        if (commandStore.isOpen()) { // then CommandBus is opened, too
+            commandBus.close();
+        }
+        eventBus.close();
+    }
+
+    /**
+     * A sample command handler that tells whether a handler was invoked.
+     */
+    class CreateProjectHandler extends CommandHandler {
+
+        private boolean handlerInvoked = false;
+
+        CreateProjectHandler(String id) {
+            super(id, eventBus);
+        }
+
+        @Assign
+        ProjectCreated handle(CreateProject command, CommandContext ctx) {
+            handlerInvoked = true;
+            return ProjectCreated.getDefaultInstance();
+        }
+
+        boolean wasHandlerInvoked() {
+            return handlerInvoked;
+        }
+    }
+}
